api/requirements.txt
Flask
mysql-python
Flask-SQLAlchemy
SQLAlchemy
gunicorn
requests
voluptuous
Pillow
markdown2
onetimepass
pyqrcode
python-memcached
nose
python-slugify